[#60] 코드 정리하기
#60
