+++ Codive/Features/Home/Presentation/View/HomeView.swift
@@ -63,20 +63,15 @@ struct HomeView: View {
             }
             .task {
                 await viewModel.loadWeather(for: nil)
-                // 오늘의 코디 여부 확인 (추가)
-//                viewModel.fetchTodayCodiData()
                 if !viewModel.isEditingExistingCodi {
-                        viewModel.fetchTodayCodiData()
-                    }
-//                await viewModel.loadActiveCategoriesWithAPI()
+                    viewModel.fetchTodayCodiData()
+                }
             }
             .onChange(of: navigationRouter.currentDestination) { newDestination in
                 if newDestination == nil {
-//                    viewModel.loadActiveCategories()
                     if viewModel.todayCodiPreview == nil {
-                                viewModel.loadActiveCategories()
-                            }
-                    // 홈으로 돌아올 때 스크롤 초기화
+                        viewModel.loadActiveCategories()
+                    }
                     scrollViewID = UUID()
                 }
             }
