+++ Codive/Features/Search/Data/Services/SearchAPIService.swift
@@ -166,7 +166,6 @@ extension SearchAPIService {
     }
 }
 
-
 // MARK: - Search API Error
 
 enum SearchAPIError: LocalizedError {
